// apps/packages/app/src/api/files/v2/google/type.ts
import { FileItem } from 'src/stores/files';
import { DriveType } from 'src/utils/interface/files';

export interface DriveMenuType {
	label: string;
	key: string | number;
	icon: string;
	driveType: DriveType;
}

export class DriveFileItem extends FileItem {
	iconLink?: string;
	webContentLink?: string;
	webViewLink?: string;
	thumbnailLink?: string;
}

export class GoogleDriveFileItem extends DriveFileItem {
	id_path: string;
	// o_path: string;
	google_file_id: string;
}